# docs/api/openapi-b.yaml
...
openapi: "3.1.0"
info: 
  title: "ai-shop B 端 BFF"
  version: "0.1.0"
  license: 
    name: "UNLICENSED"
  description: "由 b-app/src/api/endpoints.ts + api/requests.ts + packages/shared/types 自动生成，请勿手改。\n生成命令：cd b-app && npm run gen:api\n\n口径：响应包 {code,msg,data}，分页 {records,total,page,size}，camelCase，单号 xxxNo，时间 xxxAt（UTC 毫秒），枚举大写下划线，金额为最小货币单位整数。"
servers: 
  - url: "http://localhost:8080"
    description: "本地后端"
components: 
  securitySchemes: 
    bearerAuth: 
      type: "http"
      scheme: "bearer"
  schemas: 
...
    Goods: 
      type: "object"
      properties: 
        goodsNo: 
          type: "string"
          description: "商品单号"
        title: 
          type: "string"
          description: "商品标题"
        subtitle: 
          type: "string"
          description: "副标题/卖点一句话"
        cover: 
          type: "string"
          description: "封面图 URL。列表页用这一张"
        images: 
          type: "array"
          items: 
            type: "string"
          description: "详情轮播图 URL 列表"
        type: 
          "$ref": "#/components/schemas/CategoryType"
          description: "商品形态，与所属类目的 type 一致。决定详情页用哪套字段"
        categoryNo: 
          type: "string"
          description: "所属类目"
        merchant: 
          "$ref": "#/components/schemas/MerchantBrief"
          description: "所属商家 —— 商品与服务都要展示商家信息"
        rating: 
          type: "number"
          description: "本商品的评分与评价数（区别于商家整体评分）"
        ratingCount: 
          type: "number"
          description: "本商品的评价条数"
        price: 
          type: "number"
          description: "展示价（最小货币单位），取各 SKU 最低价"
        originPrice: 
          type: "number"
          description: "划线价（最小货币单位）"
        fulfillments: 
          type: "array"
          items: 
            "$ref": "#/components/schemas/FulfillmentType"
          description: "支持的履约方式。**数组**：同一商品可以既自提又快递，下单时由用户选"
        specGroups: 
          type: "array"
          items: 
            "$ref": "#/components/schemas/SpecGroup"
          description: "规格维度定义；单规格商品也有一组"
        skus: 
          type: "array"
          items: 
            "$ref": "#/components/schemas/Sku"
          description: "SKU 列表。单规格商品也有且仅有一条"
        sales: 
          type: "number"
          description: "累计销量，展示用"
        cutoffAt: 
          type: "number"
          description: "FRESH：预售截单时间戳"
        arrivalDesc: 
          type: "string"
          description: "FRESH：预计到货描述"
        weighed: 
          type: "boolean"
          description: "FRESH：是否按实称多退少补"
        origin: 
          type: "string"
          description: "FRESH：产地"
        durationMin: 
          type: "number"
          description: "SERVICE：服务时长（分钟）"
        storeName: 
          type: "string"
          description: "SERVICE：可核销门店"
        slots: 
          type: "array"
          items: 
            "$ref": "#/components/schemas/AppointmentSlot"
          description: "SERVICE + APPOINTMENT：可预约时段。**后端未下发**"
        card: 
          "$ref": "#/components/schemas/CardSpec"
          description: "CARD。**后端未下发**"
        virtual: 
          "$ref": "#/components/schemas/VirtualSpec"
          description: "VIRTUAL。**后端未下发**"
        promotions: 
          type: "array"
          items: 
            "$ref": "#/components/schemas/Promotion"
          description: "促销（一期只有买 N 送 M）。**后端未下发**"
        groupBuy: 
          type: "object"
          properties: 
            minCount: 
              type: "number"
            price: 
              type: "number"
          required: 
            - "minCount"
            - "price"
          additionalProperties: false
          description: "商家为本商品开放的拼团档：够 minCount 人享 price。不配则本商品不能发起团"
        points: 
          type: "number"
          description: "本商品每件赠送的积分。**后端未下发**：库里有 `prd_goods.points_config` 这一列， 但全仓没有任何读写。等积分域接上再兑现。"
        limitPerUser: 
          type: "number"
          description: "每人限购，0 = 不限"
        onSale: 
          type: "boolean"
          description: "是否在售。下架后详情页仍可访问（历史订单要点得进去），但不可下单"
        detail: 
          type: "string"
          description: "图文详情正文（纯文本）。空 = 商家没写 —— 端上整段不渲染， 别拿一个空白区块占着详情页。"
        status: 
          "$ref": "#/components/schemas/GoodsStatus"
        auditReason: 
          type: "string"
          description: "最近一次驳回 / 平台强制下架的原因（**只在商家侧与运营端下发，C 端恒空**）。\n\n**没有它，商家面对 `REJECTED` 只能猜要改什么** —— 审计日志只有运营看得到。 平台强制下架时后端会带「平台强制下架」前缀，商家据此知道是自己被驳 还是被平台下的。过审时清空。\n\n⚠️ 后端 `GoodsVO` 一直在发它，`MerchantGoodsService` 的注释甚至写着 「它会出现在商家 B 端（`auditReason`）」—— 而端上从没声明这个字段。 那句注释描述的是一件**从未发生过**的事。"
        titleI18n: 
          "$ref": "#/components/schemas/Record_string_string"
          description: "三语标题原文，**只有商家侧 `/biz/goods/{no}` 下发**。\n\n编辑页按语言逐格填，而保存是整份覆盖 —— 拿不到原文就只能回填当前那一格， 于是用中文改一次，英文与阿语就被清空了。**这个故障不报错**： C 端缺译文时回落中文，看起来一切正常。"
        subtitleI18n: 
          "$ref": "#/components/schemas/Record_string_string"
          description: "三语副标题原文，同 `titleI18n`"
        stdNo: 
          type: "string"
          description: "引用的平台标准品；空 = 自建品。**只有商家侧与运营端下发，C 端恒空。**\n\n<p>必须下发：编辑页保存是整份覆盖，拿不到它就等于\n**打开编辑页再保存一次就自动脱离了标准品** —— 商品从此不再被收敛， 而界面上没有任何变化。与 `titleI18n` / `priceByMarket` 是同一个形状的故障。"
      required: 
        - "goodsNo"
        - "title"
        - "subtitle"
        - "cover"
        - "images"
        - "type"
        - "categoryNo"
        - "merchant"
        - "price"
        - "fulfillments"
        - "specGroups"
        - "skus"
        - "sales"
        - "limitPerUser"
        - "onSale"
      additionalProperties: false
    GoodsStatus: 
      type: "string"
      enum: 
        - "DRAFT"
        - "ON_SALE"
        - "OFF_SALE"
        - "PENDING"
        - "REJECTED"
      description: "商家侧商品状态。\n\n<p><b>DRAFT 与 PENDING 是两件事</b>：草稿是「还没提交，等你」，待审是「已提交，等平台」—— 说错了商家的下一步就错了。也与 OFF_SALE（点一下就能卖）分开。"
...
    ToggleGoodsReq: 
      type: "object"
      properties: 
        onSale: 
          type: "boolean"
          description: "目标状态：true 上架、false 下架。下架后详情页仍可访问但不可下单"
      required: 
        - "onSale"
      additionalProperties: false
    SaveStockReq: 
      type: "object"
      properties: 
        skuNo: 
          type: "string"
          description: "要改库存的 SKU"
        stock: 
          type: "number"
          description: "改后的库存数。**是绝对值不是增量**"
      required: 
        - "skuNo"
        - "stock"
      additionalProperties: false
...
    RecognizeGoodsReq: 
      type: "object"
      properties: 
        imageUrl: 
          type: "string"
          description: "待识别的商品图 URL（先走 upload/image 拿到）。返回识别出的标题与类目建议"
      required: 
        - "imageUrl"
      additionalProperties: false
...
    Record_string_string: 
      type: "object"
      additionalProperties: 
        type: "string"
...
paths: 
...
  "/biz/goods": 
    get: 
      operationId: "mGoodsList"
      summary: "商品列表"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "page"
          in: "query"
          required: false
          schema: 
            type: "number"
            description: "页码，从 1 起"
        - name: "size"
          in: "query"
          required: false
          schema: 
            type: "number"
            description: "每页条数"
        - name: "status"
          in: "query"
          required: false
          schema: 
            anyOf: 
              - "$ref": "#/components/schemas/GoodsStatus"
              - type: "string"
                const: "OUT_OF_STOCK"
            description: "按商品状态过滤。B 端能看到全部状态，C 端只看得到 ON_SALE。\n\n<p>`OUT_OF_STOCK` **不是 `GoodsStatus` 的一员**：库里没有这个状态列， 它按「所有 SKU 可用量都 ≤ 0」算出来，且与在售**不互斥** —— 一件在售商品照样能全规格断货。B-4.1 一直写着这一筛，代码里此前没有。"
        - name: "keyword"
          in: "query"
          required: false
          schema: 
            type: "string"
            description: "按标题模糊搜。**服务层一直支持，端点此前写死传 null** —— 于是商品页没有搜索，而商品一多这一页就只能靠滚。"
        - name: "categoryNo"
          in: "query"
          required: false
          schema: 
            type: "string"
            description: "按三级类目筛。与 `keyword` 是同一种遗漏：服务层一直支持， 端点写死传 null。类目变必填之后，按类目找货是商家的主路径。"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    type: "object"
                    properties: 
                      records: 
                        type: "array"
                        items: 
                          "$ref": "#/components/schemas/Goods"
                      total: 
                        type: "integer"
                      page: 
                        type: "integer"
                      size: 
                        type: "integer"
                    required: 
                      - "records"
                      - "total"
                      - "page"
                      - "size"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/goods/{goodsNo}": 
    get: 
      operationId: "mGoodsDetail"
      summary: "商品详情"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/goods/save": 
    post: 
      operationId: "mSaveGoods"
      summary: "新建/编辑商品"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
      requestBody: 
        required: true
        content: 
          "application/json": 
            schema: 
              "$ref": "#/components/schemas/SaveGoodsReqBody"
  "/biz/goods/{goodsNo}/toggle": 
    post: 
      operationId: "mToggleGoods"
      summary: "上下架"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
      requestBody: 
        required: true
        content: 
          "application/json": 
            schema: 
              "$ref": "#/components/schemas/ToggleGoodsReq"
  "/biz/goods/{goodsNo}/stock": 
    post: 
      operationId: "mSaveStock"
      summary: "改库存"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
      requestBody: 
        required: true
        content: 
          "application/json": 
            schema: 
              "$ref": "#/components/schemas/SaveStockReq"
  "/biz/goods/{goodsNo}/store-stock": 
    post: 
      operationId: "mSaveStoreStock"
      summary: "改当前门店库存"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/goods/{goodsNo}/store-price": 
    post: 
      operationId: "mSaveStorePrice"
      summary: "改当前门店售价"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/goods/{goodsNo}/submit": 
    post: 
      operationId: "mSubmitGoods"
      summary: "提交审核（草稿→待审）"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/goods/{goodsNo}/presale": 
    post: 
      operationId: "mSavePresale"
      summary: "改截单与到货说明"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "goodsNo"
          in: "path"
          required: true
          schema: 
            type: "string"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/Goods"
                required: 
                  - "code"
                  - "msg"
                  - "data"
...
  "/biz/goods/recognize": 
    post: 
      operationId: "mRecognizeGoods"
      summary: "拍照识别商品"
      tags: 
        - "goods"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/GoodsGuess"
                required: 
                  - "code"
                  - "msg"
                  - "data"
      requestBody: 
        required: true
        content: 
          "application/json": 
            schema: 
              "$ref": "#/components/schemas/RecognizeGoodsReq"
  "/biz/category/tree": 
    get: 
      operationId: "mCategoryTree"
      summary: "类目树（选类目）"
      tags: 
        - "category"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    type: "array"
                    items: 
                      "$ref": "#/components/schemas/Category"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/spu-std": 
    get: 
      operationId: "mSpuStdSearch"
      summary: "标准品搜索（建品用）"
      tags: 
        - "spu-std"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    type: "array"
                    items: 
                      "$ref": "#/components/schemas/SpuStd"
                required: 
                  - "code"
                  - "msg"
                  - "data"
  "/biz/spec-templates": 
    get: 
      operationId: "mSpecTemplates"
      summary: "规格模板"
      tags: 
        - "spec-templates"
      security: 
        - bearerAuth: []
      parameters: 
        - name: "categoryType"
          in: "query"
          required: false
          schema: 
            type: "string"
            description: "按类目过滤平台模板；不传则返回全部 + 商家自存"
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    type: "array"
                    items: 
                      "$ref": "#/components/schemas/SpecTemplate"
                required: 
                  - "code"
                  - "msg"
                  - "data"
    post: 
      operationId: "mSaveSpecTemplate"
      summary: "存为常用规格"
      tags: 
        - "spec-templates"
      security: 
        - bearerAuth: []
      parameters: []
      responses: 
        "200": 
          description: "OK"
          content: 
            "application/json": 
              schema: 
                type: "object"
                properties: 
                  code: 
                    type: "integer"
                    description: "0 表示成功"
                  msg: 
                    type: "string"
                  data: 
                    "$ref": "#/components/schemas/SpecTemplate"
                required: 
                  - "code"
                  - "msg"
                  - "data"
      requestBody: 
        required: true
        content: 
          "application/json": 
            schema: 
              "$ref": "#/components/schemas/SaveSpecTemplateReq"
...